+++ js/app.js
@@ -27,8 +27,7 @@ OCA.Audioplayer.Core = {
     initialDocumentTitle: null,
     CategorySelectors: [],
     AjaxCallStatus: null,
-    Player: null,
-    canPlayMimeType: null,
+    canPlayMimeType: [],
 
     init: function () {
         OCA.Audioplayer.Core.initialDocumentTitle = document.title;
@@ -37,7 +36,6 @@ OCA.Audioplayer.Core = {
         OCA.Audioplayer.UI.ActivePlaylist = $('#activePlaylist');
         OCA.Audioplayer.UI.getAudiostreamUrl = OC.generateUrl('apps/audioplayer/getaudiostream') + '?t=';
 
-
         if (decodeURI(location.hash).length > 1) {
             OCA.Audioplayer.Core.processSearchResult();
         } else {
@@ -45,11 +43,16 @@ OCA.Audioplayer.Core = {
             OCA.Audioplayer.Backend.getUserValue('category', OCA.Audioplayer.Core.processCategoryFromPreset);
         }
 
-        OCA.Audioplayer.Core.canPlayMimeType = soundManager.html5;
-        var stream_array = ['audio/mpegurl', 'audio/x-scpls', 'application/xspf+xml'];
-        for (var s = 0; s < stream_array.length; s++) {
-            OCA.Audioplayer.Core.canPlayMimeType[stream_array[s]] = true;
-        }
+        // evaluate if browser can play the mimetypes
+        var mimeTypes = ['audio/mpeg', 'audio/mp4', 'audio/ogg', 'audio/wav', 'audio/flac', 'audio/x-aiff'];
+        var mimeTypeAudio = document.createElement('audio');
+        mimeTypes.forEach((element) => {
+            if (mimeTypeAudio.canPlayType(element)) {
+                OCA.Audioplayer.Core.canPlayMimeType.push(element);
+            }
+        });
+        // add playlist mimetypes
+        OCA.Audioplayer.Core.canPlayMimeType.push('audio/mpegurl', 'audio/x-scpls', 'application/xspf+xml');
 
         $('.toolTip').tooltip();
     },
@@ -64,37 +67,37 @@ OCA.Audioplayer.Core = {
                 }
             }
 
-            if (OCA.Audioplayer.Core.Player && document.querySelectorAll('#activePlaylist li').length > 0) {
+            if (OCA.Audioplayer.Player && document.querySelectorAll('#activePlaylist li').length > 0) {
                 var currentVolume;
                 var newVolume;
                 switch (e.key) {
                     case ' ':
                         if (document.getElementById('sm2-bar-ui').classList.contains('playing')) {
-                            OCA.Audioplayer.Core.Player.actions.pause();
+                            OCA.Audioplayer.Player.pause();
                         } else {
-                            OCA.Audioplayer.Core.Player.actions.resume();
+                            OCA.Audioplayer.Player.play();
                         }
                         e.preventDefault();
                         break;
                     case 'ArrowRight':
-                        OCA.Audioplayer.Core.Player.actions.next();
+                        OCA.Audioplayer.Player.next();
                         break;
                     case 'ArrowLeft':
-                        OCA.Audioplayer.Core.Player.actions.prev();
+                        OCA.Audioplayer.Player.prev();
                         break;
                     case 'ArrowUp':
-                        currentVolume = OCA.Audioplayer.Core.Player.actions.getVolume();
+                        currentVolume = OCA.Audioplayer.Player.getVolume();
                         if (currentVolume < 100) {
                             newVolume = Math.min(currentVolume + 10, 100);
-                            OCA.Audioplayer.Core.Player.actions.setVolume(newVolume);
+                            OCA.Audioplayer.Player.setVolume(newVolume);
                         }
                         e.preventDefault();
                         break;
                     case 'ArrowDown':
-                        currentVolume = OCA.Audioplayer.Core.Player.actions.getVolume();
+                        currentVolume = OCA.Audioplayer.Player.getVolume();
                         if (currentVolume > 0) {
                             newVolume = Math.max(currentVolume - 10, 0);
-                            OCA.Audioplayer.Core.Player.actions.setVolume(newVolume);
+                            OCA.Audioplayer.Player.setVolume(newVolume);
                         }
                         e.preventDefault();
                         break;
@@ -109,10 +112,8 @@ OCA.Audioplayer.Core = {
 
         document.getElementById('searchresults').classList.add('hidden');
         window.location.href = '#';
-        if (locHashTemp[0] !== 'volume' && locHashTemp[0] !== 'repeat' && locHashTemp[0] !== 'shuffle' && locHashTemp[0] !== 'prev' && locHashTemp[0] !== 'play' && locHashTemp[0] !== 'next') {
             OCA.Audioplayer.Core.CategorySelectors = locHashTemp;
             OCA.Audioplayer.Core.processCategoryFromPreset();
-        }
     },
 
     processCategoryFromPreset: function () {
@@ -134,6 +135,10 @@ OCA.Audioplayer.Core = {
                     var item = $('#individual-playlist li[data-trackid="' + OCA.Audioplayer.Core.CategorySelectors[2] + '"]');
                     item.find('.icon').hide();
                     item.find('.ioc').removeClass('ioc-volume-up').addClass('ioc-volume-off').show();
+                    if (OCA.Audioplayer.Core.CategorySelectors[3]) {
+                        // if the title was previously played, the last position will be set
+                        OCA.Audioplayer.Player.trackStartPosition = OCA.Audioplayer.Core.CategorySelectors[3];
+                    }
                 }
             });
         }
@@ -580,10 +585,11 @@ OCA.Audioplayer.UI = {
         var spanTitle = document.createElement('span');
         spanTitle.classList.add('title');
 
-        if (canPlayMimeType[elem.mim]) {
+        if (canPlayMimeType.includes(elem.mim)) {
             spanTitle.innerText = elem.cl1;
         } else {
             spanTitle.innerHTML = '<i>' + elem.cl1 + '</i>';
+            li.dataset.canPlayMime = false;
         }
 
         if (covers) {
@@ -686,49 +692,66 @@ OCA.Audioplayer.UI = {
             OCA.Audioplayer.Backend.setStatistics();
             return;
         }
-        if (!canPlayMimeType[activeLi.dataset.mimetype]) {
+        if (!canPlayMimeType.includes(activeLi.dataset.mimetype)) {
             console.warn(`can't play ${activeLi.dataset.mimetype}`);
             return false;
         }
         if (activeLi.classList.contains('isActive')) {
-            if (document.getElementById('sm2-bar-ui').classList.contains('playing')) {
-                OCA.Audioplayer.Core.Player.actions.stop();
-            } else {
-                OCA.Audioplayer.Core.Player.actions.play();
-            }
+            OCA.Audioplayer.Player.play();
         } else {
-            // the visible playlist has to be copied to the player queue
-            // this disconnects the free navigation in AP while continuing to play a playlist
-            if (document.getElementById('playlist-container').dataset.playlist !== document.getElementById('activePlaylist').dataset.playlist) {
-                var ClonePlaylist = $(playlist).clone();
-                OCA.Audioplayer.UI.ActivePlaylist.html('');
-                OCA.Audioplayer.UI.ActivePlaylist.append(ClonePlaylist);
-                OCA.Audioplayer.UI.ActivePlaylist.find('span').remove();
-                document.getElementById('activePlaylist').dataset.playlist = document.getElementById('playlist-container').dataset.playlist;
+            if (document.getElementById('playlist-container').dataset.playlist !== OCA.Audioplayer.Player.currentPlaylist) {
+                var playlistItems = document.querySelectorAll('.albumwrapper li');
+                OCA.Audioplayer.Player.addTracksToSourceList(playlistItems);
+                OCA.Audioplayer.Player.currentPlaylist = document.getElementById('playlist-container').dataset.playlist;
             }
-            OCA.Audioplayer.UI.currentTrackUiChange(coverUrl, activeLi);
-            if (OCA.Audioplayer.Core.Player.playlistController.data.selectedIndex === null) {
-                OCA.Audioplayer.Core.Player.playlistController.data.selectedIndex = 0;
-            }
+
+            OCA.Audioplayer.UI.currentTrackIndexUiChange(coverUrl, activeLi);
             var k = 0, e = activeLi;
             while (e = e.previousSibling) {
                 ++k;
             }
-            OCA.Audioplayer.Core.Player.actions.play(k);
+            // when a new title is played, the old playtime will be reset
+            if (parseInt(OCA.Audioplayer.Core.CategorySelectors[2]) !== parseInt(activeLi.dataset.trackid)) OCA.Audioplayer.Player.trackStartPosition = 0;
+            OCA.Audioplayer.Player.currentTrackIndex = k;
+            OCA.Audioplayer.Player.play();
             OCA.Audioplayer.Backend.setStatistics();
         }
     },
 
     indicateCurrentPlayingTrack: function () {
-        if (document.getElementById('playlist-container').dataset.playlist === document.getElementById('activePlaylist').dataset.playlist) {
-            var playingTrackId = document.querySelector('#activePlaylist li.selected').dataset.trackid;
-            var playingListItem = document.querySelector('.albumwrapper li[data-trackid="' + playingTrackId + '"]');
-            playingListItem.classList.add('isActive');
-            var icon = playingListItem.querySelector('.ioc');
-            icon.classList.remove('ioc-volume-off');
-            icon.classList.add('ioc-volume-up');
-            icon.style.display = 'block';
-            playingListItem.querySelector('.icon').style.display = 'none';
+        if (document.getElementById('playlist-container').dataset.playlist === OCA.Audioplayer.Player.currentPlaylist) {
+
+            if (document.getElementsByClassName('isActive').length === 1) {
+                var currentActive = document.getElementsByClassName('isActive')[0];
+                // does not work yet, when a song is preselected bot not isActive
+                //currentActive.querySelector('i.ioc').style.display = 'none';
+                //currentActive.querySelector('i.icon').style.display = 'block';
+                document.getElementsByClassName('isActive')[0].classList.remove('isActive');
+            }
+
+            // reset all playing icons
+            var iocIcon = document.querySelectorAll('.albumwrapper li i.ioc')
+            for (var i = 0; i < iocIcon.length; ++i) {
+                iocIcon[i].style.display = 'none';
+            }
+            var iconIcon = document.querySelectorAll('.albumwrapper li i.icon')
+            for (var i = 0; i < iconIcon.length; ++i) {
+                iconIcon[i].style.display = 'block';
+            }
+
+            if (!OCA.Audioplayer.Player.isPaused()) {
+                iocIcon[OCA.Audioplayer.Player.currentTrackIndex].classList.remove('ioc-volume-off');
+                iocIcon[OCA.Audioplayer.Player.currentTrackIndex].classList.add('ioc-volume-up');
+            } else {
+                iocIcon[OCA.Audioplayer.Player.currentTrackIndex].classList.add('ioc-volume-off');
+                iocIcon[OCA.Audioplayer.Player.currentTrackIndex].classList.remove('ioc-volume-up');
+            }
+            iocIcon[OCA.Audioplayer.Player.currentTrackIndex].style.display = 'block';
+            iconIcon[OCA.Audioplayer.Player.currentTrackIndex].style.display = 'none';
+
+            var nowPlayingTitle = iocIcon[OCA.Audioplayer.Player.currentTrackIndex].parentElement.parentElement.dataset.title;
+            document.getElementById('nowPlayingText').innerHTML = nowPlayingTitle;
+            document.querySelectorAll('.albumwrapper li')[OCA.Audioplayer.Player.currentTrackIndex].classList.add('isActive');
         }
     },
 
@@ -750,7 +773,7 @@ OCA.Audioplayer.UI = {
         }
     },
 
-    currentTrackUiChange: function (coverUrl, activeLi) {
+    currentTrackIndexUiChange: function (coverUrl, activeLi) {
         var addCss;
         var addDescr;
         var coverID = activeLi.dataset.cover;
@@ -761,6 +784,16 @@ OCA.Audioplayer.UI = {
             addCss = 'background-image:url(' + coverUrl + coverID + ');-webkit-background-size:cover;-moz-background-size:cover;background-size:cover;';
             addDescr = '';
         }
+        if ('mediaSession' in navigator) {
+            navigator.mediaSession.metadata = new MediaMetadata({
+                title: activeLi.dataset.title,
+                artist: activeLi.dataset.artist,
+                album: activeLi.dataset.album,
+                artwork: [
+                    {src: coverUrl + coverID, sizes: '192x192', type: 'image/png'},
+                ]
+            });
+        }
         document.querySelector('.sm2-playlist-cover').setAttribute('style', addCss);
         document.querySelector('.sm2-playlist-cover').innerText = addDescr;
         document.title = activeLi.dataset.title + ' (' + activeLi.dataset.artist + ' ) @ ' + OCA.Audioplayer.Core.initialDocumentTitle;
@@ -768,9 +801,9 @@ OCA.Audioplayer.UI = {
 
     soundmanagerCallback: function (SMaction) {
         if (SMaction === 'setVolume') {
-            OCA.Audioplayer.Backend.setUserValue('volume', Math.round(OCA.Audioplayer.Core.Player.actions.getVolume()));
+            OCA.Audioplayer.Backend.setUserValue('volume', Math.round(OCA.Audioplayer.Player.getVolume()));
         } else {
-            OCA.Audioplayer.UI.currentTrackUiChange(
+            OCA.Audioplayer.UI.currentTrackIndexUiChange(
                 OC.generateUrl('apps/audioplayer/getcover/'),
                 document.querySelector('#activePlaylist li.selected')
             );
@@ -815,24 +848,25 @@ OCA.Audioplayer.UI = {
         });
         $('#individual-playlist').append(elems.slice(0));
 
-        if (document.getElementById('playlist-container').dataset.playlist === document.getElementById('activePlaylist').dataset.playlist) {
+        if (document.getElementById('playlist-container').dataset.playlist === OCA.Audioplayer.Player.currentPlaylist) {
             OCA.Audioplayer.UI.ActivePlaylist.append(elems);
         }
 
-        if (OCA.Audioplayer.Core.Player) {
+        if (OCA.Audioplayer.Player) {
             var e = document.querySelector('#activePlaylist li.selected');
             if (e) {
                 var k = 0;
                 while (e = e.previousSibling) {
                     ++k;
                 }
-                OCA.Audioplayer.Core.Player.playlistController.data.selectedIndex = k;
+                OCA.Audioplayer.Player.playlistController.data.selectedIndex = k;
             }
         }
     },
 
     resizePlaylist: function () {
         document.getElementById('sm2-bar-ui').style.width = document.getElementById('playlist-container').offsetWidth + 'px';
+        document.getElementById('progressBar').width = document.getElementById('progressContainer').offsetWidth;
         if (document.querySelector('.is-active')) {
             if (document.getElementById('playlist-container').offsetWidth < 850) {
                 document.querySelector('.songcontainer-cover').classList.add('cover-small');
@@ -868,7 +902,85 @@ OCA.Audioplayer.UI = {
             }
             return false;
         }
+    },
+
+    whatsNewSuccess: function (data, statusText, xhr) {
+        if (xhr.status !== 200) {
+            return
+        }
+
+        let item, menuItem, text, icon
+
+        const div = document.createElement('div')
+        div.classList.add('popovermenu', 'open', 'whatsNewPopover', 'menu-left')
+
+        const list = document.createElement('ul')
+
+        // header
+        item = document.createElement('li')
+        menuItem = document.createElement('span')
+        menuItem.className = 'menuitem'
+
+        text = document.createElement('span')
+        text.innerText = t('core', 'New in') + ' ' + data['product']
+        text.className = 'caption'
+        menuItem.appendChild(text)
+
+        icon = document.createElement('span')
+        icon.className = 'icon-close'
+        icon.onclick = function () {
+            OCA.Audioplayer.Backend.whatsnewDismiss(data['version'])
+        }
+        menuItem.appendChild(icon)
+
+        item.appendChild(menuItem)
+        list.appendChild(item)
+
+        // Highlights
+        for (const i in data['whatsNew']['regular']) {
+            const whatsNewTextItem = data['whatsNew']['regular'][i]
+            item = document.createElement('li')
+
+            menuItem = document.createElement('span')
+            menuItem.className = 'menuitem'
+
+            icon = document.createElement('span')
+            icon.className = 'icon-checkmark'
+            menuItem.appendChild(icon)
+
+            text = document.createElement('p')
+            text.innerHTML = _.escape(whatsNewTextItem)
+            menuItem.appendChild(text)
+
+            item.appendChild(menuItem)
+            list.appendChild(item)
+        }
+
+        // Changelog URL
+        if (!_.isUndefined(data['changelogURL'])) {
+            item = document.createElement('li')
+
+            menuItem = document.createElement('a')
+            menuItem.href = data['changelogURL']
+            menuItem.rel = 'noreferrer noopener'
+            menuItem.target = '_blank'
+
+            icon = document.createElement('span')
+            icon.className = 'icon-link'
+            menuItem.appendChild(icon)
+
+            text = document.createElement('span')
+            text.innerText = t('core', 'View changelog')
+            menuItem.appendChild(text)
+
+            item.appendChild(menuItem)
+            list.appendChild(item)
+        }
+
+        div.appendChild(list)
+        document.body.appendChild(div)
     }
+
 };
 
 /**
@@ -922,7 +1034,7 @@ OCA.Audioplayer.Backend = {
     },
 
     setStatistics: function () {
-        var track_id = $('#activePlaylist li.selected').data('trackid');
+        var track_id = OCA.Audioplayer.Player.currentTrackId;
         if (track_id) {
             $.ajax({
                 type: 'GET',
@@ -948,6 +1060,26 @@ OCA.Audioplayer.Backend = {
         });
     },
 
+    whatsnew: function (options) {
+        options = options || {}
+        $.ajax({
+            type: 'GET',
+            url: OC.generateUrl('apps/audioplayer/whatsnew'),
+            data: {'format': 'json'},
+            success: options.success || function (data, statusText, xhr) {
+                OCA.Audioplayer.UI.whatsNewSuccess(data, statusText, xhr)
+            },
+        });
+    },
+
+    whatsnewDismiss: function dismiss(version) {
+        $.ajax({
+            type: 'POST',
+            url: OC.generateUrl('apps/audioplayer/whatsnew'),
+            data: {version: encodeURIComponent(version)}
+        })
+        $('.whatsNewPopover').remove()
+    }
 };
 
 /**
@@ -1079,7 +1211,7 @@ OCA.Audioplayer.Playlists = {
 
                 eventTarget.addClass('sortActive');
                 if (document.getElementById('sm2-bar-ui').classList.contains('playing')) {
-                    OCA.Audioplayer.Core.Player.actions.pause();
+                    OCA.Audioplayer.Player.pause();
                     $('#individual-playlist li').removeClass('isActive');
                     $('#individual-playlist li i.ioc').hide();
                 } else {
@@ -1218,10 +1350,7 @@ document.addEventListener('DOMContentLoaded', function () {
     OCA.Audioplayer.Core.initKeyListener();
     OCA.Audioplayer.Backend.checkNewTracks();
     OCA.Audioplayer.Playlists.initPlaylistActions();
-    if (document.getElementById('audioplayer_sonos').value !== 'checked') {
-        OCA.Audioplayer.Core.Player = new SM2BarPlayer(document.getElementById('sm2-bar-ui'));
-        OCA.Audioplayer.Core.Player.actions.setVolume(document.getElementById('audioplayer_volume').value);
-    }
+    OCA.Audioplayer.Backend.whatsnew();
 
     var notify = document.getElementById('audioplayer_notification').value;
     if (notify !== '') {
@@ -1265,7 +1394,8 @@ document.addEventListener('DOMContentLoaded', function () {
     document.querySelector('.header-album').addEventListener('click', OCA.Audioplayer.UI.sortPlaylist);
 
     window.setTimeout(function () {
-        document.getElementById('sm2-bar-ui').style.width = document.getElementById('playlist-container').offsetWidth + 'px'
+        document.getElementById('sm2-bar-ui').style.width = document.getElementById('playlist-container').offsetWidth + 'px';
+        document.getElementById('progressBar').width = document.getElementById('progressContainer').offsetWidth;
     }, 1000);
 
     var resizeTimeout;
@@ -1281,4 +1411,23 @@ document.addEventListener('DOMContentLoaded', function () {
             OCA.Audioplayer.Core.processSearchResult();
         }
     };
+
+    // mediaSession currently use for Chrome already to support hardware keys
+    if ('mediaSession' in navigator) {
+        navigator.mediaSession.setActionHandler('play', function () {
+            OCA.Audioplayer.Player.play();
+        });
+        navigator.mediaSession.setActionHandler('pause', function () {
+            OCA.Audioplayer.Player.pause();
+        });
+        navigator.mediaSession.setActionHandler('stop', function () {
+            OCA.Audioplayer.Player.stop();
+        });
+        navigator.mediaSession.setActionHandler('previoustrack', function () {
+            OCA.Audioplayer.Player.prev();
+        });
+        navigator.mediaSession.setActionHandler('nexttrack', function () {
+            OCA.Audioplayer.Player.next();
+        });
+    }
 });